Fixes to pip install + miscellaneous changes
- Updated MANIFEST.in
- Fix build-system when installing via pip
- Fixed AppleClang issue with hash specialization of static_string
- Moved _config.cpp to config.cpp
