+++ pyproject.toml
@@ -1,3 +1,12 @@
+[build-system]
+requires = [
+    "setuptools >= 40.0.4",
+    "setuptools_scm >= 2.0.0",
+    "wheel >= 0.29.0",
+    "scikit-build >= 0.8.0",
+    "cython",
+]
+build-backend = 'setuptools.build_meta'
 
 [tool.black]
 line-length = 80
